Use Debian 10 for the 32-bit environment
Some Synology users face a problem with glibc >= 2.31 due to collisions
in the syscall number assignments.  Synology uses some that have since
been used for other things in mainline Linux.  The one that exposed the
issue here is clock_gettime64 being #403 in mainline Linux while in the
Synology kernel exhibiting the issue it is SYNOArchiveBit.  Suffice it
to say these functions are not interchangeable.  It seems that glibc can
not tell the difference and working around the issue is required.

This seemed as clean as providing a wrapper library using LD_PRELOAD
that ensured none of the 32-bit programs this image cared about would
step on the wrong syscalls, because it is difficult to control what
programs may call what other programs.  This is admittedly a little
dirty, since the x86_64 execution will be using glibc 2.31 and the i386
environment will be using 2.28.

In practice I did not see any adverse effects, but it is not heavily
exercised.  Just in launching valheim-server with -crossplay and
ensuring libpary.so loaded successfully.