Add SERVER_PUBLIC to cron environment
When SERVER_PUBLIC is set to false the server does not respond on the
query port.  When server_is_idle() is run under cron environment
SERVER_PUBLIC is unset and defaults to true causing the function to
attempt to use the query port.

This patch adds SERVER_PUBLIC to the cron environment so that
server_is_idle() will use the correct path for determining if players
are online.

See #551 Restart if idle only restarts when NOT idle.